Created interface FormRecordable.
