+++ api/src/main/java/org/openmrs/FormRecordable.java
@@ -9,25 +9,12 @@
  */
 package org.openmrs;
 
-import org.apache.commons.lang3.StringUtils;
-import org.openmrs.api.APIException;
-
-import javax.persistence.Column;
-import javax.persistence.MappedSuperclass;
-
 /**
- * This is a super class to make a bridge between an OpenMrsObject and is position in the form.
+ * Defines the methods to make a bridge between an OpenMrsObject and is position in the form.
+ *
+ * @since 1.4
  */
-@MappedSuperclass
-public abstract class FormRecordable extends BaseChangeableOpenmrsData{
-	
-	private static final String FORM_NAMESPACE_PATH_SEPARATOR = "^";
-
-	private static final int FORM_NAMESPACE_PATH_MAX_LENGTH = 255;
-	
-	@Column(name = "form_namespace_path")
-	protected String formNamespaceAndPath;
-
+public interface FormRecordable {
 
 	/**
 	 * Gets the namespace for the form field that was used to capture the obs details in the form
@@ -38,17 +25,7 @@ public abstract class FormRecordable extends BaseChangeableOpenmrsData{
 	 * <strong>Should</strong> return the correct namespace for a form field with a path
 	 * <strong>Should</strong> return null if the namespace is not specified
 	 */
-	public String getFormFieldNamespace() {
-		if (StringUtils.isNotBlank(formNamespaceAndPath)) {
-			//Only the path was specified
-			if (formNamespaceAndPath.startsWith(FORM_NAMESPACE_PATH_SEPARATOR)) {
-				return null;
-			}
-			return formNamespaceAndPath.substring(0, formNamespaceAndPath.indexOf(FORM_NAMESPACE_PATH_SEPARATOR));
-		}
-
-		return formNamespaceAndPath;
-	}
+	String getFormFieldNamespace();
 
 	/**
 	 * Gets the path for the form field that was used to capture the obs details in the form
@@ -59,17 +36,7 @@ public abstract class FormRecordable extends BaseChangeableOpenmrsData{
 	 * <strong>Should</strong> return the correct path for a form field with a namespace
 	 * <strong>Should</strong> return null if the path is not specified
 	 */
-	public String getFormFieldPath() {
-		if (StringUtils.isNotBlank(formNamespaceAndPath)) {
-			//Only the namespace was specified
-			if (formNamespaceAndPath.endsWith(FORM_NAMESPACE_PATH_SEPARATOR)) {
-				return null;
-			}
-			return formNamespaceAndPath.substring(formNamespaceAndPath.indexOf(FORM_NAMESPACE_PATH_SEPARATOR) + 1);
-		}
-
-		return formNamespaceAndPath;
-	}
+	String getFormFieldPath();
 
 	/**
 	 * Sets the namespace and path of the form field that was used to capture the obs details in the
@@ -78,7 +45,7 @@ public abstract class FormRecordable extends BaseChangeableOpenmrsData{
 	 * form applications can subtract the length of their namespace from 254 to determine the
 	 * maximum length they can use for a form field path.
 	 *
-	 * @param namespace the namespace of the form field
+	 * @param namespace     the namespace of the form field
 	 * @param formFieldPath the path of the form field
 	 * @since 1.11
 	 * <strong>Should</strong> set the underlying formNamespaceAndPath in the correct pattern
@@ -90,32 +57,5 @@ public abstract class FormRecordable extends BaseChangeableOpenmrsData{
 	 * <strong>Should</strong> mark the obs as dirty when the value is changed from a null to a non null value
 	 * <strong>Should</strong> mark the obs as dirty when the value is changed from a non null to a null value
 	 */
-	public void setFormField(String namespace, String formFieldPath) {
-		if (namespace == null && formFieldPath == null) {
-			markAsDirty(formNamespaceAndPath, null);
-			formNamespaceAndPath = null;
-			return;
-		}
-
-		String nsAndPathTemp = "";
-		if (StringUtils.isNotBlank(namespace) && StringUtils.isNotBlank(formFieldPath)) {
-			nsAndPathTemp = namespace + FORM_NAMESPACE_PATH_SEPARATOR + formFieldPath;
-		} else if (StringUtils.isNotBlank(namespace)) {
-			nsAndPathTemp = namespace + FORM_NAMESPACE_PATH_SEPARATOR;
-		} else if (StringUtils.isNotBlank(formFieldPath)) {
-			nsAndPathTemp = FORM_NAMESPACE_PATH_SEPARATOR + formFieldPath;
-		}
-
-		if (nsAndPathTemp.length() > FORM_NAMESPACE_PATH_MAX_LENGTH) {
-			throw new APIException("Obs.namespaceAndPathTooLong", (Object[]) null);
-		}
-		if (StringUtils.countMatches(nsAndPathTemp, FORM_NAMESPACE_PATH_SEPARATOR) > 1) {
-			throw new APIException("Obs.namespaceAndPathNotContainSeparator", (Object[]) null);
-		}
-
-		markAsDirty(this.formNamespaceAndPath, nsAndPathTemp);
-		formNamespaceAndPath = nsAndPathTemp;
-	}	
-		
-	protected abstract void markAsDirty(Object oldValue, Object newValue);
+	void setFormField(String namespace, String formFieldPath);
 }
